Remove FontData::pos and always pass by value
In my initial design I had imagined that FontData would track its
position relative to the data root, so that when a parse error occured
we could store the absolute position of that error, which we could use
in debugging.

I never ended up really purusing this, and so the field was superfluous.

In addition to removing it, this moves to always passing FontData by
value; in effect FontData is a transparent wrapper over &'a [u8]. This
means that we can copy it around, and this in turn simplifies some of
our generated code.