[codegen] No repr(packed) for non-zerocopy records
The only rationale for adding these is to make the struct usable via
transmute & co, and that is not something these types are doing. Having
these annotations is potentially confusing, and also likely produces a
sub-optimal memory layout.

This also updates the signature for FontData::read_array_unchecked to
include a T: FixedSized bound; this omission was a bug.