[codegen] More recursive FromObjRef/FromTableRef
This makes these implementations less bespoke in the case of offset
fields: these are now resolved by recursively calling one of these
methods, and leaving the conversion logic to blanket impls.

This also solves an issue where we were using a bunch of weird Into
impls to handle conversion of offsets, which was about to run into
'conflicting impl' issues, as we discover more and more offset targets,
and try to figure out how to resolve them.