add pre-push git hook to check for accidental println! statements
similar to https://github.com/googlefonts/fontmake-rs/pull/204, with the only difference that in the case of fontations there are a couple of files where the use of println! is intentional and legitimate (font-codegen::preprocessor or otexplorer) so I added an allow-list to be able to exclude those from the check.