[codegen] Support specifying format positions
When we generate an enum for a 'format' table (a table for
which there are multiple possible representations, distinguished by a
format value stored at a consistent location) we were previously
assuming that the format value was always stored at the head of the
table.

I want to use the format mechanism for Device/VariationIndex tables, but
in this case the 'format' is stored four bytes into the table. This
patch adds support for specifying an offset at which the format should
be read; e.g. you can know write,

   ```
   format u16@4 MyTable {
      // etc
   }
   ```
And we will read the format from the 4th byte in the input data.