[write-fonts] Use shared points in GlyphVariations
This is a basic initial implementation, that needs a bit of fixing up;
in particular we should look into whether or not it is worth sharing the
'all' value (the zero byte) which would only be saving a byte per tuple,
and would only make sense if it is common for fonts to often have lots
of tuples that use all points? but hey, a byte is a byte.

Also I have an idea for a fancy but slow algorithm to compute more
efficient share points, and that could be fun to look into.