[read-fonts] Add 'get' method to coverage tables
I wanted this for testing the table splitting in Mark2Base, and it's
something we will want anyway.

This also fixes a little mixup in the signature of CoverageTable::iter;
the elided lifetime here meant the compiler tied the iterator to the
lifetime of the table object, whereas we really want to tie it to the
lifetime of the underlying data.