[graph] Move assert to after space check
It is totally possible for us to encounter pathological graphs that
contain overflows in the root space, which cannot be resolved.
Previously we had an assert that would fire if this happened, which was
incorrect.

This moves the assert to after we have already filtered out overflows in
the root space.

It also makes this (and a related) assert run in release as well as
debug, since this is not hot code and we expect the compiler to be run
in release even during testing, in some instances.