[write-fonts] Add Glyph::Empty
This is a cleaner way of representing the idea of a glyph with no
outlines, but which should still be added to loca.

I played around with also adding this in read-fonts, but it doesn't
really play well with codegen or parsing, there, because we expect a
format enum to have a table for each variant, and in the empty case we
have no data from which to parse a table.