Big glyph ids (#994)
Renames current `GlyphId` to `GlyphId16` and adds a new 32-bit `GlyphId`.

All codegen inputs changed to use explicit 16-bit ids.

Changes all skrifa public APIs to accept/return the new type. Some parts of read-fonts and write-fonts have been updated to work with the larger type but not all. We might need to work this out as we go.

Fixes #784